refactor: split iron proxy cli config
